Moved Inviter class into InviterList.cs
